move mcm to seed lib
