+++ seed/lib/mcm/cleaners.py
@@ -1,13 +1,13 @@
 """
 :copyright: (c) 2014 Building Energy Inc
 """
-import dateutil
 from datetime import datetime, date
 import re
 import string
 
-from mcm.matchers import fuzzy_in_set
+import dateutil
 
+from seed.lib.mcm.matchers import fuzzy_in_set
 
 NONE_SYNONYMS = (u'not available', u'not applicable', u'n/a')
 BOOL_SYNONYMS = (u'true', u'yes', u'y', u'1')
